[refactor/postsNetworkBoundResource] refactor: rename PostDao to PostDao in the database package and update imports
